+++ packages/frontend/features/Panes/CollectionItemsHeader.tsx
@@ -1,8 +1,9 @@
 import { useCallback, useRef } from 'react';
 import { useContextSelector } from 'use-context-selector';
-import { CollectionHeader } from '@components/collection/header';
+import { CollectionHeader, MenuAction } from '@components/collection/header';
 import {
   useCollectionsContext,
+  useMarkCollectionAsRead,
   useRefreshCollection,
   useSetCollectionLayout,
 } from '@features/Collections';
@@ -27,7 +28,6 @@ export const CollectionItemsHeader: React.FC = () => {
       : false;
 
   const { mutate: refreshCollection } = useRefreshCollection();
-
   const handleRefreshClick = useCallback(() => {
     if (activeCollection) {
       const collectionId = activeCollection.id;
@@ -37,7 +37,7 @@ export const CollectionItemsHeader: React.FC = () => {
 
       refreshCollection({ id: collectionId });
     } else {
-      console.error(`onRefreshClicked() without active collection`);
+      console.error(`handleRefreshClick() without active collection`);
     }
   }, [activeCollection, refreshCollection]);
 
@@ -52,14 +52,43 @@ export const CollectionItemsHeader: React.FC = () => {
     [activeCollection.id, setCollectionLayout]
   );
 
+  const { mutate: markCollectionAsRead } = useMarkCollectionAsRead();
+  const handleMarkAsRead = useCallback(() => {
+    if (activeCollection) {
+      const collectionId = activeCollection.id;
+      if (typeof collectionId === 'string' && collectionId === 'home') {
+        return;
+      }
+
+      markCollectionAsRead({ id: collectionId });
+    } else {
+      console.error(`handleMarkAsRead() without active collection`);
+    }
+  }, [activeCollection, markCollectionAsRead]);
+
+  const handleMenuActionClicked = useCallback(
+    (action: MenuAction) => {
+      switch (action) {
+        case 'refresh': {
+          handleRefreshClick();
+          break;
+        }
+        case 'markAsRead': {
+          handleMarkAsRead();
+        }
+      }
+    },
+    [handleMarkAsRead, handleRefreshClick]
+  );
+
   return (
     <CollectionHeader
       collection={activeCollection}
       isRefreshing={isRefreshing}
       menuButtonRef={drawerButtonRef}
-      onRefresh={handleRefreshClick}
-      onMenuClicked={toggleDrawer}
+      onHamburgerClicked={toggleDrawer}
       onChangeLayout={handleCollectionLayoutChanged}
+      onMenuActionClicked={handleMenuActionClicked}
     />
   );
 };